newsnow/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "newsnow-scheduler"
version = "1.0.0"
description = "新闻数据自动获取调度器 - 定时获取各新闻源数据并持久化到MongoDB"
authors = [
    {name = "NewsNow Team", email = "orongxing@gmail.com"},
]
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "httpx>=0.25.0",
    "beautifulsoup4>=4.12.0",
    "pymongo>=4.5.0",
    "apscheduler>=3.10.0",
    "feedparser>=6.0.0",
    "pydantic>=2.4.0",
    "loguru>=0.7.0",
    "python-dotenv>=1.0.0",
    "lxml>=4.9.0",
    "aiofiles>=23.0.0",
    "tenacity>=8.2.0",
    "click>=8.1.0",
    "brotli>=1.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]

[project.scripts]
newsnow = "newsnow_py.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["newsnow_py"]

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]
asyncio_mode = "auto"
